% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fam_dat.R
\docType{data}
\name{fam_dat}
\alias{fam_dat}
\title{Simulated time-to-event data.}
\format{
Data.frame.
}
\usage{
data(fam_dat)
}
\description{
This toy data set is simulated for educational purposes explaining the
package. It concerns of families in which at least two individuals
experienced the event of interest during following up. The covariate of
interest is risk modifier 'x'. This data set is inspired by data that
is often seen in genetic epidemiological studies in cancer research.
}
\references{
Refer to paper.
(\href{weblink}{PubMed})
}
\keyword{datasets}
