% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_data.R
\name{Prepare_data}
\alias{Prepare_data}
\title{Prepare outcome-dependently sampled time-to-event data for weight calculation.}
\usage{
Prepare_data(dat, population_incidence, breaks)
}
\arguments{
\item{dat}{Data.frame with one row per individual which at least includes
a column \strong{d} with event indicator (1 for event, 0 for censored), a column
\strong{y} with event/censoring time.}

\item{population_incidence}{Either a vector (in combination with breaks) or
a data.frame (columns start age group, end age group, incidence rate).}

\item{breaks}{Cut-points for the (age/time) groups.}
}
\value{
Data.frame in long format (one interval of an individual per row)
with a.o. the columns \strong{S_k} population interval-based survival; \strong{p_k}
population interval-based hazard; \strong{S_k.} sample interval-based survival;
\strong{p_k.} sample interval-based hazard.
}
\description{
This function prepares a data set to use for the calculation of
weights that correct for ascertainment bias, @seealso \code{\link[=Calculate_weights]{Calculate_weights()}}.
The specific ascertainment bias addressed is introduced by outcome-dependent
sampling of clustered time-to-events, for example by inclusion of high-risk
families in genetic epidemiological studies in cancer research.
}
\details{
In the corresponding paper the sample interval-based survival (S_k.)
is distinguished from the population equivalent with an apostroph (') instead
of a full stop (.). The same holds for the sample interval-based hazard
(p_k.).
}
