% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Prepare_data.R
\name{Prepare_data}
\alias{Prepare_data}
\title{Prepare outcome-dependently sampled time-to-event data for weight calculation.}
\usage{
Prepare_data(dat, population_incidence, breaks)
}
\arguments{
\item{dat}{Data.frame with one row per individual which at least includes
a column \strong{d} with event indicator (1 for event, 0 for censored), a column
\strong{y} with event/censoring time.}

\item{population_incidence}{A vector (in combination with breaks) or
a data.frame (columns 1) 'start age group', 2) 'end age group', 3)'S_pop') with
population incidence per 100,000 per interval k.}

\item{breaks}{Cut-points for the (age/time) groups. Only needed when population_incidence is a vector.}
}
\value{
Data.frame one row per individual and a.o. columns \emph{d}
non-censoring indicator; \strong{k} interval of (age) group; \strong{S_k}
population interval-based proportion of individuals experiencing the
event in intervals later than k; \strong{S_k.} sample
proportion of individuals experiencing the event in intervals later
than k.
}
\description{
This function prepares a data set to use for the calculation of
weights that correct for ascertainment bias, @seealso \code{\link[=Calculate_weights]{Calculate_weights()}}.
The specific ascertainment bias addressed is introduced by outcome-dependent
sampling of clustered time-to-events, for example by inclusion of high-risk
families in genetic epidemiological studies in cancer research.
}
\details{
In the corresponding paper the sample interval-based survival (S_k.)
is distinguished from the population equivalent with a hyperscript 0 instead
of a full stop (.).
}
